+++ src/main/java/com/arcanc/pulselib/content/model/animation/PAnimation.java
@@ -15,13 +15,14 @@ import org.jetbrains.annotations.Nullable;
 import org.joml.Quaternionf;
 import org.joml.Vector3f;
 
+import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 
 public record PAnimation(String name,
                          float length,
                          Map<String, PBoneAnimation> boneAnimations,
-                         List<PAnimationEvent> events)
+                         List<PAnimationEvent<?>> events)
 {
 	public PAnimation(String name, float length, Map<String, PBoneAnimation> boneAnimations)
 	{
@@ -90,11 +91,19 @@ public record PAnimation(String name,
 				scale == null ? new Vector3f(1f) : scale);
 	}
 
-	public List<PAnimationEvent> eventsBetween(float from, float to)
+	public List<PAnimationEvent<?>> eventsBetween(float from, float to)
 	{
 		if (this.events.isEmpty() || to < from)
 			return List.of();
 		return this.events.stream().filter(event -> (event.time() > from || (from == 0f && event.time() == 0f)) && event.time() <= to).toList();
+	}
+
+	public List<PAnimationEvent<?>> eventsBetweenReverse(float from, float to)
+	{
+		if (this.events.isEmpty() || to > from)
+			return List.of();
+		return this.events.stream().filter(event -> event.time() >= to && event.time() < from).
+			sorted(Comparator.comparingDouble((PAnimationEvent<?> event) -> event.time()).reversed()).toList();
 	}
 
 	private static <T> @Nullable T sample(PBoneAnimation boneAnimation,